refactor: Update routing configuration to enable guest access to the auth module
This commit modifies the routing configuration in the app-routing.module.ts file to allow guest users to access the auth module. The canActivate and canActivateChild properties for the auth module route have been uncommented, enabling the GuestGuard to control access to the module. This change ensures that only authenticated users can access the reviews and dashboard modules, while allowing guest users to access the auth module for registration and login purposes.